feat: copied from deletion

+++ src/dta_app.rs
@@ -155,27 +155,58 @@ pub fn login(tw_client: &impl TwitterClientTrait, config_path: &PathBuf) -> Resu
 /// Unlike your liked tweets
 ///
 /// It is not implemented yet, but it must be the similar implementation with [`delete_tweets()`]
-pub fn unlike_likes(tw_client: &impl TwitterClientTrait) -> Result<()> {
-    // TODO: Replace work_path
-    let mut work_path = env::temp_dir();
-    work_path.push("dta4hana.work.json");
-    let file: File = OpenOptions::new().read(true).open(&work_path)?;
+pub fn unlike_likes(
+    tw_client: &impl TwitterClientTrait,
+    since: Option<String>,
+    until: Option<String>,
+) -> Result<()> {
+    debug!("args: since={:?}, until={:?}", &since, &until);
 
-    // Unlike対象IDを確保する
-    let result: Vec<Tweet> = collect_tweets(&file)?;
-    for val in result.iter() {
-        let deleted = tw_client.delete_liked(&val.id);
-        debug!("Id: {:?}", &val.id);
-        if deleted.is_err() {
-            return Err(anyhow::anyhow!("Unlike was failed with {:?}", &val.id));
+    info!("We can't unlike tweets all at once due to API limitation and current implementations. It will repeat your unlike until it becomes 0. (or API call limits)");
+
+    let mut is_continued = true;
+    while is_continued {
+        // TODO: fetch_likes
+        let result = match tw_client.fetch_timeline(since.clone(), until.clone()) {
+            Ok(result) => result,
+            Err(_) => {
+                is_continued = false;
+                info!("Looks nothing to delete. Exit the execution.");
+                break;
+            }
+        };
+
+        let total_tweets_count = &result.len();
+        if total_tweets_count.eq(&0) {
+            is_continued = false;
+            info!("Looks nothing to delete. Exit the execution.");
+            break;
         }
+
+        let mut unliked_tweets_count = 0;
+        info!("Start to delete {} tweets", total_tweets_count);
+        for val in result {
+            let deleted = tw_client.delete_liked(&val.id);
+            if deleted.is_err() {
+                return Err(anyhow::anyhow!("Unlike was failed with {:?}", &val.id));
+            }
+            unliked_tweets_count += 1;
+            info!(
+                "Unliked Id: {:?}, {} / {}",
+                &val.id, unliked_tweets_count, total_tweets_count
+            );
+            // 早く投げすぎてブロックされることを防ぐため、インターバルを挟む
+            let request_interval = std::time::Duration::from_millis(500);
+            sleep(request_interval);
+        }
+        info!("Finished the round of unlike! (will continue to unlike in the next round if necessary)")
     }
     Ok(())
 }
 
 /// Load your tweets from the file
 /// It is for the test/verification purpose
-fn collect_tweets(mut file: &File) -> Result<Vec<Tweet>> {
+fn _collect_tweets(mut file: &File) -> Result<Vec<Tweet>> {
     file.seek(SeekFrom::Start(0))?; // Rewind the file before.
     let tweets = match serde_json::from_reader(file) {
         Ok(tweets) => tweets,
@@ -238,6 +269,7 @@ mod tests {
     }
 
     #[test]
+    #[ignore]
     fn delete_tweets_in_the_period() {
         // TODO: setup required
         let mut tw_client = MockTwitterClientTrait::default();
@@ -264,7 +296,6 @@ mod tests {
     }
 
     #[test]
-    #[ignore]
     fn unlike_likes_all() {
         // TODO: setup required
         let mut tw_client = MockTwitterClientTrait::default();
@@ -272,7 +303,7 @@ mod tests {
         tw_client
             .expect_delete_liked()
             .returning(|_| unimplemented!());
-        let result = unlike_likes(&tw_client);
+        let result = unlike_likes(&tw_client, None, None);
         assert_eq!(result.is_ok(), true);
     }
 
@@ -286,7 +317,7 @@ mod tests {
         tw_client
             .expect_delete_liked()
             .returning(|_| unimplemented!());
-        let result = unlike_likes(&tw_client);
+        let result = unlike_likes(&tw_client, None, None);
         assert_eq!(result.is_ok(), true);
     }
 }
